Momoco script
Remove once per duel summon with once per turn. 
Add condition on (1) that it's effects are negated until end phase. This will give opponent's chance to counter momoco as it will be able to be returned and targeted by card effects in this short frame as well as more use to momoco overall.
(5) to ignition.
(9) effect; Cannot be destroyed by battle if you control 3 or more monster's with different types. Set banlist limit to 1.